--- app/src/App.tsx
@@ -1,13 +1,0 @@
-import React from 'react';
-import './App.css';
-import { ImageDrop } from './components/ImageDrop/ImageDrop'
-
-function App() {
-  return (
-    <div className="App">
-      <ImageDrop />
-    </div>
-  );
-}
-
-export default App;
