--- app/src/decs.d.ts
@@ -1,2 +1,0 @@
-declare module 'get-pixels';
-declare module 'quantize';
