+++ src/App.tsx
@@ -1,0 +1,12 @@
+import './App.css';
+import { ImageDrop } from './components/ImageDrop/ImageDrop'
+
+function App() {
+  return (
+    <div className="App">
+      <ImageDrop />
+    </div>
+  );
+}
+
+export default App;
