+++ src/app/[[...slug]]/page.tsx
@@ -1,0 +1,10 @@
+'use client';
+
+import dynamic from 'next/dynamic';
+import '../../index.css';
+
+const App = dynamic(() => import('../../App'), { ssr: false });
+
+export default function Page() {
+  return <App />;
+}
